Fix incorrect image being generated

+++ JazziconSwift-Demo/JazziconSwift-Demo/ViewController.swift
@@ -35,7 +35,7 @@ class ViewController: UIViewController {
     
     func generate() {
         let seed = UInt64(seedTextField.text ?? "")
-        let image = Jazzicon(seed: seed ?? 1).generateImage(size: 200)
+        let image = Jazzicon(seed: UInt32(seed ?? 1)).generateImage(size: 200)
         imageView.image = image
     }
 }
